\name{Rweights-package}
\alias{Rweights-package}
\alias{Rweights}
\docType{package}
\title{
Functions for weighting survey data
}
\description{
User provides target proportions, variables, and data. The result is a survey design object and vector of weights. Can be used for iterative raking.
}
\details{
\tabular{ll}{
Package: \tab Rweights\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-10-30\cr
License: \tab  GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Loren Collingwood
Maintainer: <loren.collingwood@ucr.edu>
}

\references{
Thomas Lumley, ``Survey Analysis in R'' <http://faculty.washington.edu/tlumley/survey/>

}

\keyword{ package }
\seealso{
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
